third_party/blink/web_tests/http/tests/security/xss-DENIED-window-name-navigator-expected.txt

CONSOLE MESSAGE: iframe1: SecurityError: Failed to read a named property 'navigator' from 'Window': Blocked a frame with origin "null" from accessing a cross-origin frame.
CONSOLE MESSAGE: iframe1: SecurityError: Failed to read a named property 'navigator' from 'Window': Blocked a frame with origin "null" from accessing a cross-origin frame.
CONSOLE MESSAGE: iframe2: SecurityError: Failed to read a named property 'navigator' from 'Window': Blocked a frame with origin "null" from accessing a cross-origin frame.
 
